Fix incorrect test
